+++ luxonis_train/loaders/__init__.py
@@ -1,12 +1,10 @@
-from .base_loader import BaseLoaderTorch
+from .base_loader import BaseLoaderTorch, LuxonisLoaderTorchOutput
 from .luxonis_loader_torch import LuxonisLoaderTorch
 from .luxonis_perlin_loader_torch import LuxonisLoaderPerlinNoise
-from .utils import LuxonisLoaderTorchOutput, collate_fn
 
 __all__ = [
     "BaseLoaderTorch",
     "LuxonisLoaderPerlinNoise",
     "LuxonisLoaderTorch",
     "LuxonisLoaderTorchOutput",
-    "collate_fn",
 ]
